Revert "Revert "sp2. 모임 키워드화 배포""
