+++ package.json
@@ -31,7 +31,7 @@
     "@sopt-makers/fonts": "^2.0.2",
     "@sopt-makers/icons": "^1.1.0",
     "@sopt-makers/playground-common": "^1.7.2",
-    "@sopt-makers/ui": "^2.8.8",
+    "@sopt-makers/ui": "^2.8.12",
     "@stitches/react": "^1.2.8",
     "@tanstack/react-query": "^4.10.3",
     "@types/autosize": "^4.0.3",
