+++ pages/detail/index.tsx
@@ -1,9 +1,9 @@
+import { GetMeetingResponse } from '@api/API_LEGACY/meeting';
 import { useQueryGetMeeting } from '@api/API_LEGACY/meeting/hooks';
-import { useRouter } from 'next/router';
+import CommonDetail from '@components/page/detail';
 import dayjs from 'dayjs';
 import 'dayjs/locale/ko';
-import CommonDetail from '@components/page/detail';
-import { GetMeetingResponse } from '@api/API_LEGACY/meeting';
+import { useRouter } from 'next/router';
 
 dayjs.locale('ko');
 
@@ -11,6 +11,8 @@ const DetailPage = () => {
   const router = useRouter();
   const id = router.query.id as string;
   const { data: meetingData } = useQueryGetMeeting({ params: { id } });
+
+  if (meetingData?.category === '번쩍') router.replace(`/detail/flash?id=${id}`);
 
   return <CommonDetail detailData={meetingData as GetMeetingResponse} />;
 };
