+++ src/api/flash/index.ts
@@ -10,9 +10,6 @@ export const createFlash = async (formData: FlashFormType) => {
 };
 
 const filterFlashFormData = (formData: FlashFormType) => {
-  const convertedTags = formData.welcomeTags?.map(tag => {
-    return tag?.value;
-  });
   const convertedEndDate =
     formData.timeInfo.time.value === '당일' ? formData.timeInfo.dateRange[0] : formData.timeInfo.dateRange[1];
   const convertedFlashPlace = formData.placeInfo.place.value === '협의 후 결정' ? null : formData.placeInfo.placeDetail;
@@ -29,7 +26,8 @@ const filterFlashFormData = (formData: FlashFormType) => {
       maximumCapacity: formData.capacityInfo.maxCapacity,
       files: formData.files,
     },
-    welcomeMessageTypes: convertedTags?.length === 0 ? null : convertedTags,
+    meetingKeywordTypes: formData.meetingKeywordTypes,
+    welcomeMessageTypes: formData.welcomeMessageTypes,
   };
   return data;
 };
