+++ src/components/page/list/Card/DesktopSizeCard/index.tsx
@@ -1,30 +1,34 @@
+import { MeetingListResponse } from '@api/API_LEGACY/meeting';
+import ProfileDefaultIcon from '@assets/svg/profile_default.svg?rect';
+import { CategoryChip } from '@components/page/list/Card/DesktopSizeCard/CategoryChip';
+import { MeetingInformation } from '@components/page/list/Card/DesktopSizeCard/constant';
+import RecruitmentStatusTag from '@components/Tag/RecruitmentStatusTag';
 import { Flex } from '@components/util/layout/Flex';
 import { CategoryKoType } from '@constants/option';
-import { MeetingListOfFilterResponse } from '@api/API_LEGACY/meeting';
-import { styled } from 'stitches.config';
-import ProfileDefaultIcon from '@assets/svg/profile_default.svg?rect';
+import { Tag } from '@sopt-makers/ui';
 import { getResizedImage } from '@utils/image';
-import { CategoryChip } from '@components/page/list/Card/DesktopSizeCard/CategoryChip';
-import RecruitmentStatusTag from '@components/Tag/RecruitmentStatusTag';
-import { MeetingInformation } from '@components/page/list/Card/DesktopSizeCard/constant';
+import { styled } from 'stitches.config';
 
 interface CardProps {
-  meetingData: MeetingListOfFilterResponse['meetings'][number];
+  meetingData: MeetingListResponse['meetings'][number];
   isFlash?: boolean;
-  welcomeMessageTypes?: string[];
   flashDetailInfo?: {
     label: string;
     value: () => string;
   }[];
+  flashCount?: string;
 }
 
-function DesktopSizeCard({ meetingData, isFlash = false, welcomeMessageTypes, flashDetailInfo }: CardProps) {
+function DesktopSizeCard({ meetingData, isFlash = false, flashDetailInfo, flashCount }: CardProps) {
   const detailInfo = isFlash && flashDetailInfo ? flashDetailInfo : MeetingInformation(meetingData);
 
   return (
     <div>
       <ImageWrapper>
         <RecruitmentStatusTag status={meetingData.status} style={{ position: 'absolute', top: '16px', left: '16px' }} />
+        <STag size="md" type="solid">
+          {isFlash ? flashCount : `${meetingData.approvedCount} / ${meetingData.capacity}명`}
+        </STag>
         <SThumbnailImage
           css={{
             backgroundImage: `url(${getResizedImage(meetingData.imageURL[0]?.url ?? '', 380)})`,
@@ -33,7 +37,10 @@ function DesktopSizeCard({ meetingData, isFlash = false, welcomeMessageTypes, fl
       </ImageWrapper>
 
       <STitleSection>
-        <CategoryChip category={meetingData.category as CategoryKoType} welcomeMessage={welcomeMessageTypes} />
+        <CategoryChip
+          category={meetingData.category as CategoryKoType}
+          meetingKeywordTypes={meetingData.meetingKeywordTypes}
+        />
         <STitle>{meetingData.title}</STitle>
       </STitleSection>
 
@@ -61,6 +68,13 @@ export default DesktopSizeCard;
 const ImageWrapper = styled('div', {
   position: 'relative',
 });
+
+const STag = styled(Tag, {
+  position: 'absolute',
+  top: '16px',
+  right: '16px',
+});
+
 const SThumbnailImage = styled('div', {
   width: '380px',
   height: '260px',
