+++ src/components/page/list/Card/MobileSize/index.tsx
@@ -1,17 +1,16 @@
-import { MeetingListOfFilterResponse } from '@api/API_LEGACY/meeting';
+import { MeetingListResponse } from '@api/API_LEGACY/meeting';
 import CardType from './CardType';
 import ListType from './ListType';
 
 export interface MobileSizeCardProps {
-  meetingData: MeetingListOfFilterResponse['meetings'][0];
-  isAllParts: boolean;
+  meetingData: MeetingListResponse['meetings'][0];
   mobileType: 'list' | 'card';
 }
 
-function MobileSizeCard({ meetingData, isAllParts, mobileType }: MobileSizeCardProps) {
+function MobileSizeCard({ meetingData, mobileType }: Omit<MobileSizeCardProps, 'isAllParts'>) {
   switch (mobileType) {
     case 'list':
-      return <ListType meetingData={meetingData} isAllParts={isAllParts} />;
+      return <ListType meetingData={meetingData} />;
     case 'card':
     default:
       return <CardType meetingData={meetingData} />;
