+++ src/components/page/list/Filter/DropDown/index.tsx
@@ -1,75 +1,79 @@
 import { ampli } from '@/ampli';
 import { FilterType } from '@constants/option';
-import { useQueryString } from '@hooks/queryString';
-import { IconXCircle } from '@sopt-makers/icons';
+import { useMultiQueryString } from '@hooks/queryString';
+import useDebounce from '@hooks/useDebounce';
 import { SelectV2 } from '@sopt-makers/ui';
 import { css } from '@stitches/react';
-import { useRouter } from 'next/router';
-import React from 'react';
+import { useEffect, useMemo, useState } from 'react';
 import { styled } from 'stitches.config';
 
 interface DropDownFilterProps {
   filter: FilterType;
+  width?: string;
 }
 
 //notice: 현재 클래스 제대로 적용안되는 문제점으로 인해 !important 사용 필요
-const autoClass = css({
-  width: '120px !important',
-  whiteSpace: 'nowrap',
-});
+const getAutoClass = (width?: string) =>
+  css({
+    minWidth: `${width ?? '140px'} !important`,
+    width: `${width ?? '140px'} !important`,
+    whiteSpace: 'nowrap',
+  });
 
-function DropDownFilter({ filter }: DropDownFilterProps) {
+function DropDownFilter({ filter, width }: DropDownFilterProps) {
   const { subject, options, label } = filter;
-  const { value: selectedValue, setValue, deleteKey } = useQueryString(subject);
+  const { value: selectedValue, setValue, deleteKey } = useMultiQueryString(subject);
+  const [rawSelected, setRawSelected] = useState<string[]>([]);
+  const debounceValue = useDebounce(rawSelected, 1300);
+  const defaultValue = useMemo(() => selectedValue.map((opt: string) => ({ label: opt, value: opt })), [selectedValue]);
 
-  const router = useRouter();
-  const selectedPartQuery = router.query[subject] as string;
+  const resolvedLabel = useMemo(() => {
+    const selected = rawSelected.length > 0 ? rawSelected : selectedValue;
+    // 단일 선택시 label 은 유저가 선택한 값
+    if (selected.length === 1) return selected[0];
+    // 다중 선택 시 label 은 필터라벨
+    else return label;
+  }, [label, rawSelected, selectedValue]);
 
-  const isActiveGeneration = subject === 'isOnlyActiveGeneration' && selectedPartQuery === 'true';
-  const defaultValue = isActiveGeneration
-    ? { label: '36기', value: '36기' }
-    : selectedPartQuery
-    ? { label: selectedPartQuery, value: selectedPartQuery }
-    : undefined;
-
-  const setPartQuery = (value: string | null) => {
-    if (value === null) {
+  const setPartQuery = (value: string | string[]) => {
+    const values = typeof value === 'string' ? [value] : value;
+    if (values.length === 0) {
+      setRawSelected([]);
       return deleteKey();
     }
-
-    //notice: 활동 기수 드롭다운의 경우, 특별 처리
-    if (subject === 'isOnlyActiveGeneration' && value === '36기') {
-      if (selectedValue) return deleteKey();
-      setValue('true');
-      ampli.clickFilterGeneration({ group_generation: true });
-      return;
-    }
-
-    ampli.clickFilterPart({ group_part: value });
-
-    if (selectedValue === value) return deleteKey();
-    return setValue(value);
+    setRawSelected(values);
   };
+
+  const handleAmpliLog = (value: string[]) => {
+    const joined = value.join(',');
+
+    switch (subject) {
+      case 'category':
+        ampli.clickFilterCategory({ group_category: joined });
+        break;
+      case 'status':
+        ampli.clickFilterStatus({ group_status: joined });
+        break;
+      case 'part':
+        ampli.clickFilterPart({ group_part: joined });
+        break;
+      default:
+        break;
+    }
+  };
+
+  useEffect(() => {
+    if (debounceValue && rawSelected?.length > 0) {
+      setValue(debounceValue);
+      handleAmpliLog(debounceValue);
+    }
+  }, [debounceValue]);
 
   return (
     <SDropDownContainer>
-      <SelectV2.Root type="text" visibleOptions={6} defaultValue={defaultValue} onChange={setPartQuery}>
+      <SelectV2.Root type="text" visibleOptions={6} defaultValue={defaultValue} onChange={setPartQuery} multiple={true}>
         <SelectV2.Trigger>
-          <SelectV2.TriggerContent
-            className={autoClass()}
-            placeholder={label}
-            icon={
-              defaultValue ? (
-                <IconXCircle
-                  style={{ width: '20px', height: '20px', fill: 'white', color: 'black' }}
-                  onClick={e => {
-                    e.stopPropagation();
-                    setPartQuery(null);
-                  }}
-                />
-              ) : null
-            }
-          />
+          <SelectV2.TriggerContent className={getAutoClass(width)()} placeholder={label} label={resolvedLabel} />
         </SelectV2.Trigger>
         <SelectV2.Menu>
           {options.map(option => (
