+++ src/components/page/list/Filter/index.tsx
@@ -2,9 +2,10 @@ import { styled } from 'stitches.config';
 import { Flex } from '@components/util/layout/Flex';
 import { useSearchParams } from '@hooks/queryString/custom';
 import Search from './Search';
-import { CATEGORY_FILTER, GENERATION_FILTER, PART_FILTER, STATUS_FILTER } from '@constants/option';
+import { CATEGORY_FILTER, GENERATION_FILTER, KEYWORD_FILTER, PART_FILTER, STATUS_FILTER } from '@constants/option';
 import DropDownFilter from './DropDown';
 import { useDisplay } from '@hooks/useDisplay';
+import FilterResetButton from '@components/page/list/Filter/Reset';
 
 function Filter() {
   const { value: search } = useSearchParams();
@@ -16,10 +17,12 @@ function Filter() {
       <Flex align="center" justify="between">
         <ScrollFilter>
           <Flex>
-            <DropDownFilter filter={CATEGORY_FILTER} />
-            <DropDownFilter filter={PART_FILTER} />
-            <DropDownFilter filter={STATUS_FILTER} />
-            <DropDownFilter filter={GENERATION_FILTER} />
+            <DropDownFilter filter={CATEGORY_FILTER} width={'149px'} />
+            <DropDownFilter filter={KEYWORD_FILTER} width={'133px'} />
+            <DropDownFilter filter={STATUS_FILTER} width={'152px'} />
+            <DropDownFilter filter={PART_FILTER} width={'121px'} />
+            <DropDownFilter filter={GENERATION_FILTER} width={'110px'} />
+            <FilterResetButton />
           </Flex>
         </ScrollFilter>
