+++ src/components/page/list/Grid/List.tsx
@@ -1,5 +1,6 @@
 import { ampli } from '@/ampli';
 import { useGetMeetingAds } from '@api/advertisement/hook';
+import { MeetingListResponse } from '@api/API_LEGACY/meeting';
 import { useQueryMeetingListOfAll } from '@api/API_LEGACY/meeting/hooks';
 import { useQueryMeetingListOfApplied, useQueryMeetingListOfMine, useQueryMyProfile } from '@api/API_LEGACY/user/hooks';
 import { usePageParams } from '@hooks/queryString/custom';
@@ -104,7 +105,8 @@ export function MeetingListOfMine() {
           {mineData?.meetings.map(meetingData => (
             <Card
               key={meetingData.id}
-              meetingData={meetingData}
+              // TODO: mine meetingData 에 welcomeMessageTypes, meetingKeywordTypes 가 현재 없지만, 곧 서버에서 내려줄 예정
+              meetingData={meetingData as unknown as MeetingListResponse['meetings'][number]}
               mobileType="card"
               bottom={
                 meetingData?.isCoLeader ? (
@@ -135,7 +137,8 @@ export function MeetingListOfApplied() {
           {applyData?.apply.map(applyData => (
             <Card
               key={applyData.id}
-              meetingData={applyData.meeting}
+              // TODO: mine meetingData 에 welcomeMessageTypes, meetingKeywordTypes 가 현재 없지만, 곧 서버에서 내려줄 예정
+              meetingData={applyData.meeting as unknown as MeetingListResponse['meetings'][number]}
               mobileType="card"
               bottom={<Status status={applyData.status} />}
             />
