+++ src/types/form.ts
@@ -1,6 +1,6 @@
 import dayjs from 'dayjs';
-import { z } from 'zod';
 import customParseFormat from 'dayjs/plugin/customParseFormat';
+import { z } from 'zod';
 dayjs.extend(customParseFormat);
 
 const capacitySchema = z.number({
@@ -32,6 +32,11 @@ export const schema = z.object({
       invalid_type_error: '카테고리를 선택해주세요.',
     }),
   }),
+  meetingKeywordTypes: z
+    .array(z.string())
+    .max(2, { message: '최대 2개까지 선택할 수 있어요' })
+    .min(1, { message: '키워드를 선택해주세요' }),
+  welcomeMessageTypes: z.array(z.string()).max(3, { message: '최대 3개까지 선택할 수 있어요' }).optional().nullable(),
   files: z
     .array(z.string(), { required_error: '이미지를 추가해주세요.' })
     .min(1, { message: '이미지를 추가해주세요.' }),
@@ -273,18 +278,11 @@ export const flashSchema = z.object({
       }
     }),
   files: z.array(z.string()),
-  welcomeTags: z
-    .array(
-      z
-        .object({
-          label: z.string(),
-          value: z.string(),
-        })
-        .optional()
-        .nullable()
-    )
-    .optional()
-    .nullable(),
+  welcomeMessageTypes: z.array(z.string()).max(3, { message: '최대 3개까지 선택할 수 있어요' }).optional().nullable(),
+  meetingKeywordTypes: z
+    .array(z.string())
+    .max(2, { message: '최대 2개까지 선택할 수 있어요' })
+    .min(1, { message: '키워드를 선택해주세요' }),
 });
 
 export type FlashFormType = z.infer<typeof flashSchema>;
